In this case test_CI.py minimize_scalar fails
Returns 0.32 rather than 0.42 or similar. minimize with Nelder-Mead
works. Output:

Distance for Y = 0.0 0.5932223141033222
Distance for Y = 0.04473684210526316 0.5932223141033222
Distance for Y = 0.08947368421052632 0.7062155919638684
Distance for Y = 0.13421052631578947 0.11299327786054619
Distance for Y = 0.17894736842105263 1.0704597667477755
Distance for Y = 0.2236842105263158 0.9574664888872293
Distance for Y = 0.26842105263157895 1.0610547513662194
Distance for Y = 0.3131578947368421 0.10358826247899011
Distance for Y = 0.35789473684210527 0.14479557197735282
Distance for Y = 0.40263157894736845 0.04120730949836271
Distance for Y = 0.4473684210526316 0.04120730949836271
Distance for Y = 0.4921052631578947 0.574470133159039
Distance for Y = 0.5368421052631579 0.5332628236606763
Distance for Y = 0.5815789473684211 0.9203069926528219
Distance for Y = 0.6263157894736842 0.38704416899214555
Distance for Y = 0.6710526315789473 0.42955016166460935
Distance for Y = 0.7157894736842105 0.0425059926724638
Distance for Y = 0.7605263157894737 0.05538306605381482
Distance for Y = 0.8052631578947369 0.012877073381351023
Distance for Y = 0.85 0.012877073381351023
     fun: 0.10358826247899011
 message: 'Solution found.'
    nfev: 14
     nit: 14
  status: 0
 success: True
       x: 0.32420863452106696
Returning Ybest = 0.32420863452106696